enable heap sizes up to at 64GB

+++ rtglobals.c
@@ -39,9 +39,9 @@ BPTR first_partition_ptr;
 BPTR last_partition_ptr;
 
 BPTR write_vector;
-int write_vector_size;
+size_t write_vector_size;
 
-int total_partition_pages;
+long total_partition_pages;
 int unmarked_color;
 int marked_color;
 int enable_write_barrier;
